# [d3d11] Don't perform extraneous queue flushes on NVIDIA