--- a/src/d3d11/d3d11_context_imm.cpp
+++ b/src/d3d11/d3d11_context_imm.cpp
@@ -737,7 +737,12 @@ namespace dxvk {
     uint64_t sequenceNumber = GetCurrentSequenceNumber();
     pResource->TrackSequenceNumber(Subresource, sequenceNumber);
 
-    FlushImplicit(TRUE);
+    // Don't flush early on NV
+    if (!m_device->adapter()->matchesDriver(DxvkGpuVendor::Nvidia,
+                                            VK_DRIVER_ID_NVIDIA_PROPRIETARY,
+                                            0,
+                                            0))
+      FlushImplicit(TRUE);
   }
 
 
@@ -746,7 +751,12 @@ namespace dxvk {
     uint64_t sequenceNumber = GetCurrentSequenceNumber();
     pResource->TrackSequenceNumber(sequenceNumber);
 
-    FlushImplicit(TRUE);
+    // Don't flush early on NV
+    if (!m_device->adapter()->matchesDriver(DxvkGpuVendor::Nvidia,
+                                            VK_DRIVER_ID_NVIDIA_PROPRIETARY,
+                                            0,
+                                            0))
+      FlushImplicit(TRUE);
   }