@K0bin 2022-03-17
If this commit causes such a large performance hit, that's probably not exclusive to Nvidia. As far as I understand it, the commit was basically just made speculative, so it's probably a better idea to just revert it entirely.