@Joshua-Ashton 2022-03-17
Queue flushing like this is not really a vendor specific action, so ye, should see the impact on radv too