@liam-middlebrook 2022-03-18
I don't have any hardware that's supported by RADV, so I can't confirm there; but I'm happy to change the commit to just be a revert of https://github.com/doitsujin/dxvk/commit/34fd16b8f22af023abe23cb68daafc23d6f23bbe if that's the direction people want.