@doitsujin 2022-03-18
How many queue submissions per frame are we talking about here?

Reverting that commit in question seems like a rather bad idea since it can lead unstable performance in games that are limited by GPU readbacks.