@doitsujin 2022-03-19
The real game seems to be sitting at a handful of submissions per frame, which seems reasonable:
![Bildschirmfoto](https://user-images.githubusercontent.com/25567304/159134863-53c7c267-b73c-4223-8f9d-e94ab3e607aa.png)

Does the capture replay thing you're using insert a lot of staging buffer copies or something?