@liam-middlebrook 2022-03-22
I tried to launch the full game, but kept hitting UPlay login issues.

Anyways, the HUD stats from my run show similar submission count, my change knocks out a single submission.

Before:
![2022-03-21-185206_3840x2160_scrot](https://user-images.githubusercontent.com/3920942/159415387-4a2f0cf8-de9a-470b-927c-d40afbeee9be.png)

After:
![2022-03-21-185221_3840x2160_scrot](https://user-images.githubusercontent.com/3920942/159415393-5ffb03a5-0955-474b-84c9-ff16fc1cf1d8.png)

> I guess we can implement the stall detection logic we have in place for queries though to reduce flushes.
That seems reasonable; out of curiosity how often are you seeing games have a usage-pattern that would be impacted by that?