@K0bin 2022-03-22
Do you know where those 2 queue syncs are coming from?