@doitsujin 2023-01-17
`master` completely reworked queue submission logic, so this is no longer useful.